backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
supabase==2.4.0

sqlalchemy==2.0.25

# Async tasks
celery==5.3.6
redis==5.0.1

# LLM and AI
openai==1.10.0
tiktoken==0.5.2

# Vector database
chromadb==0.4.22
numpy<2.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.27.2
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
sentence-transformers==2.2.2
langchain-community==0.0.1
chromadb==0.4.22
groq==0.7.2